% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Abline}
\alias{Abline}
\title{Abline Function}
\usage{
Abline(name, ...)
}
\arguments{
\item{name}{(character) The name of the parameter.}

\item{...}{Additional arguments to be passed.}
}
\value{
(NULL)
}
\description{
Description:
This function creates a new global parameter for controlling line properties in a plot.
}
\examples{
Abline("myline", alp.abline = 0.5, col.abline = "blue", lwd.abline = 2, sld.abline = FALSE)

}
\references{
For more information, refer to the documentation of the \code{newGlobalParam} function.
}
\seealso{
\code{\link{newGlobalParam}}
}
\keyword{global}
\keyword{line}
\keyword{parameter}
\keyword{plot}
