% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{color.class}
\alias{color.class}
\title{Color Class Identifier}
\usage{
color.class()
}
\value{
The identifier for the color class.
}
\description{
The \code{color.class} function returns the identifier for the color class.
}
\examples{
color.class() # Returns "color.000000IM9fKDwnkPL8A0dxaV3zT64Q6cjKifaT0BA9kmcvCAQ1NT"

}
