% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color.combine}
\alias{color.combine}
\title{Color Combine}
\usage{
color.combine(..., r, abs = F)
}
\arguments{
\item{...}{Colors to combine, specified as color objects.}

\item{r}{Numeric vector or scalar representing the weights for combining the colors.}

\item{abs}{Logical value indicating whether the color distances should be absolute (TRUE) or not (FALSE).
Default is FALSE.}
}
\value{
A color object representing the combined color.
}
\description{
This function combines multiple colors based on the specified weights, creating a new color.
}
\examples{
# Combine two colors with equal weights
col1 <- color(r = 0.2, g = 0.4, b = 0.6)
col2 <- color(r = 0.8, g = 0.2, b = 0.4)
combined_col <- color.combine(col1, col2, r = 0.5)
combined_col$getRGB() # Get the RGB values of the combined color
combined_col$getCol() # Get the combined color as an R color object

# Combine three colors with different weights
col3 <- color(r = 0.1, g = 0.5, b = 0.3)
combined_col2 <- color.combine(col1, col2, col3, r = c(0.3, 0.5, 0.2))
combined_col2$getRGB() # Get the RGB values of the combined color
combined_col2$getCol() # Get the combined color as an R color object

}
