% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color.rainbow}
\alias{color.rainbow}
\title{Rainbow Colors}
\usage{
color.rainbow(n, abs = F, inv = F)
}
\arguments{
\item{n}{Number of colors in the rainbow sequence.}

\item{abs}{Logical value indicating whether the color distances should be absolute (TRUE) or not (FALSE).
Default is FALSE.}

\item{inv}{Logical value indicating whether to invert the sequence of colors. If TRUE, the sequence will be inverted.
Default is FALSE.}
}
\value{
A color object representing the rainbow sequence of colors.
}
\description{
This function generates a sequence of rainbow colors.
}
\examples{
# Generate a rainbow sequence of colors with 6 colors
rainbow_col <- color.rainbow(n = 6)
rainbow_col$getRGBs() # Get the RGB values of the rainbow color sequence
rainbow_col$getCols() # Get the rainbow color sequence as a matrix of R color objects

# Generate an inverted rainbow sequence of colors with 8 colors
inverted_rainbow_col <- color.rainbow(n = 8, inv = TRUE)
inverted_rainbow_col$getRGBs() # Get the RGB values of the inverted rainbow color sequence
inverted_rainbow_col$getCols() # Get the inverted rainbow color sequence as a matrix of R color objects

}
