% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\title{Color}
\usage{
color(r = NA, g = NA, b = NA, maxValue = 1)
}
\arguments{
\item{r}{Numeric vector or character name representing the red component of the color.
If a character name is provided, it will be converted into an RGB value.}

\item{g}{Numeric vector or character name representing the green component of the color.
If a character name is provided, it will be converted into an RGB value.}

\item{b}{Numeric vector or character name representing the blue component of the color.
If a character name is provided, it will be converted into an RGB value.}

\item{maxValue}{The maximum value for the RGB components. Default is 1.}
}
\value{
A color object with various functions to manipulate and combine colors.
}
\description{
This function creates a color object based on the provided RGB values. The RGB values
can be specified as numeric vectors or character names of colors. If the RGB values are
not provided, random values are used by default.
}
\examples{
# Create a color object with specific RGB values
col1 <- color(r = 0.2, g = 0.4, b = 0.6)
col1$getRGB() # Get the RGB values of the color
col1$getCol() # Get the color as an R color object

# Create a color object from a character name
col2 <- color("red")
col2$getRGB() # Get the RGB values of the color
col2$getCol() # Get the color as an R color object

# Combine two colors
combined_col <- col1$combine(col2, r = 0.7)
combined_col$getRGB() # Get the RGB values of the combined color
combined_col$getCol() # Get the combined color as an R color object

}
