% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color.sequence}
\alias{color.sequence}
\title{Color Sequence}
\usage{
color.sequence(..., n, abs = F, inv = F)
}
\arguments{
\item{...}{Colors to combine, specified as color objects.}

\item{n}{Number of colors in the sequence.}

\item{abs}{Logical value indicating whether the color distances should be absolute (TRUE) or not (FALSE).
Default is FALSE.}

\item{inv}{Logical value indicating whether to invert the sequence of colors. If TRUE, the sequence will be inverted.
Default is FALSE.}
}
\value{
A color object representing the sequence of colors.
}
\description{
This function generates a sequence of colors by combining multiple colors based on the specified weights.
}
\examples{
# Generate a sequence of colors using two base colors
col1 <- color(r = 0.2, g = 0.4, b = 0.6)
col2 <- color(r = 0.8, g = 0.2, b = 0.4)
sequence_col <- color.sequence(col1, col2, n = 5)
sequence_col$getRGBs() # Get the RGB values of the color sequence
sequence_col$getCols() # Get the color sequence as a matrix of R color objects

# Generate an inverted sequence of colors
inverted_col <- color.sequence(col1, col2, n = 4, inv = TRUE)
inverted_col$getRGBs() # Get the RGB values of the inverted color sequence
inverted_col$getCols() # Get the inverted color sequence as a matrix of R color objects

}
