% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create.dir}
\alias{create.dir}
\title{Create a directory and its parent directories if they don't exist}
\usage{
create.dir(path)
}
\arguments{
\item{path}{The path of the directory to be created.}
}
\value{
The function does not return any value.
}
\description{
The \code{create.dir} function is used to create a directory in the file system and its parent directories if they don't exist. If the specified directory already exists, no action is taken.
}
\examples{
create.dir("path/to/directory")

}
