% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{csv.printer}
\alias{csv.printer}
\title{Print data from a list of CSV scans}
\usage{
csv.printer(lista, basis = NULL)
}
\arguments{
\item{lista}{The list object containing the scanned data.}

\item{basis}{An optional basis string to prepend to each file path. If not provided, the first scanned file path is used as the basis.}
}
\description{
The \code{csv.printer} function is used to print data from a list of CSV scans. It creates directories and writes CSV files based on the scanned data.
}
\examples{
csv.printer(lista)

}
