% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{csv.scanner}
\alias{csv.scanner}
\title{Read and scan a CSV file}
\usage{
csv.scanner(file, lista = NULL, basis = "")
}
\arguments{
\item{file}{The path to the CSV file.}

\item{lista}{An optional list object to append the scanned data to. If not provided, a new list is created.}

\item{basis}{A basis string to prepend to each scanned line. This can be useful for adding a prefix to the scanned data.}
}
\value{
A list object containing the scanned data.
}
\description{
The \code{csv.scanner} function is used to read and scan a CSV file. It returns the scanned data as a list.
}
\examples{
csv.scanner("path/to/file.csv")

}
