% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{fixCol}
\alias{fixCol}
\title{Fix Color}
\usage{
fixCol(col)
}
\arguments{
\item{col}{The color argument to fix. It can be NULL, a character, or a color object.}
}
\value{
A valid color object.
}
\description{
This function fixes the color argument by converting it into a valid color object.
If the color argument is NULL, it returns a default color. If the color argument is
a character, it converts it into a color object. If the color argument is already a
valid color object, it returns it as is.
}
