% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{funList}
\alias{funList}
\title{Apply a function to each element of a list}
\usage{
funList(fun, x, ...)
}
\arguments{
\item{fun}{The function to apply to each element of the list.}

\item{x}{The input list.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A new list with the results of applying the function to each element of \code{x}.
}
\description{
This function applies a specified function to each element of a list and returns
a new list with the results.
}
\examples{
# Define a function to calculate the square of a number
square <- function(x) x^2

# Apply the square function to each element of a list
lst <- list(1, 2, 3, 4, 5)
result <- funList(square, lst)
result # Returns [1, 4, 9, 16, 25]

}
