% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{funVec}
\alias{funVec}
\title{Apply a function to each element of a vector}
\usage{
funVec(fun, x, ...)
}
\arguments{
\item{fun}{The function to apply to each element of the vector.}

\item{x}{The input vector.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A new vector with the results of applying the function to each element of \code{x}.
}
\description{
This function applies a specified function to each element of a vector and
returns a new vector with the results.
}
\examples{
# Define a function to calculate the square of a number
square <- function(x) x^2

# Apply the square function to each element of a vector
vec <- c(1, 2, 3, 4, 5)
result <- funVec(square, vec)
result # Returns [1, 4, 9, 16, 25]

}
