% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getSeconds}
\alias{getSeconds}
\title{Get Seconds}
\usage{
getSeconds()
}
\value{
The current system time in seconds.
}
\description{
The \code{getSeconds} function is used to retrieve the current system time in seconds.
}
\examples{
getSeconds() # Retrieves the current system time in seconds

}
