% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{globalVar}
\alias{globalVar}
\title{Set or retrieve a global variable}
\usage{
globalVar(name, ...)
}
\arguments{
\item{name}{The name of the global variable.}

\item{...}{Optional value to set the global variable. If not provided,
retrieves the current value of the global variable.}
}
\value{
If no value is provided, returns the current value of the global variable.
If a value is provided, sets the global variable to that value.
}
\description{
This function allows you to set or retrieve a global variable. Global variables
are stored in a persistent list, which is stored as an internal object.
}
\examples{
# Set a global variable
globalVar("count", 10)

# Retrieve the value of the global variable
globalVar("count")
# Returns: 10

# Update the value of the global variable
globalVar("count", 15)

# Retrieve the updated value of the global variable
globalVar("count")
# Returns: 15

}
