% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{group.class}
\alias{group.class}
\title{Group Class Identifier}
\usage{
group.class()
}
\value{
The identifier for the group class.
}
\description{
The \code{group.class} function returns the identifier for the group class.
}
\examples{
group.class() # Returns "group.00000004intEgVdJq9HfQwRwDXWRlZYxDSX646IH6s51zkcT1B"

}
