% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Hist}
\alias{Hist}
\title{Hist Function}
\usage{
Hist(name, ...)
}
\arguments{
\item{name}{(character) The name of the parameter.}

\item{...}{Additional arguments to be passed.}
}
\value{
(NULL)
}
\description{
Description:
This function creates a new global parameter for controlling histogram plot properties.
}
\examples{
Hist("myhist", relative = TRUE, horizontal = TRUE, new.plot = TRUE, alp.hist = 0.7,
     col.hist = NULL, lty.hist = 1, lwd.hist = 2)

}
\references{
For more information, refer to the documentation of the \code{newGlobalParam} function.
}
\seealso{
\code{\link{newGlobalParam}}
}
\keyword{global}
\keyword{histogram}
\keyword{parameter}
