% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{IF}
\alias{IF}
\title{Conditional Function}
\usage{
IF(cond, x, y = NULL)
}
\arguments{
\item{cond}{The condition to evaluate.}

\item{x}{The value to return if the condition is \code{TRUE}.}

\item{y}{The value to return if the condition is \code{FALSE}. Defaults to \code{NULL}.}
}
\value{
The value of \code{x} if the condition is \code{TRUE}, or the value of \code{y} if the condition is \code{FALSE}.
}
\description{
The \code{IF} function is a convenient wrapper around the \code{if} statement in R. It allows you to evaluate a condition and return different values based on the outcome.
}
\examples{
IF(TRUE, "Condition is true") # Returns "Condition is true"
IF(FALSE, "Condition is true") # Returns NULL
IF(FALSE, "Condition is true", "Condition is false") # Returns "Condition is false"

}
