% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interval}
\alias{interval}
\title{Calculate confidence interval for a numeric vector or matrix.}
\usage{
interval(x, level)
}
\arguments{
\item{x}{The input numeric vector or matrix.}

\item{level}{The confidence level, expressed as a percentage (e.g., 95 for a 95\% confidence level).}
}
\value{
If the input is a matrix, a list of confidence intervals for each column. If the input is a vector, a list with 'inf' and 'sup' indicating the lower and upper bounds of the confidence interval.
}
\description{
This function calculates the confidence interval for a given numeric vector or matrix. If the input is a matrix, it returns a list of confidence intervals for each column of the matrix. If the input is a vector, it returns the confidence interval for the vector.
}
\examples{
data <- c(1, 2, 3, 4, 5)
interval(data, 95) # Calculate the 95\% confidence interval for the vector
matrix <- matrix(1:10, ncol = 2)
interval(matrix, 90) # Calculate the 90\% confidence interval for each column of the matrix
}
