% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.clonable}
\alias{is.clonable}
\title{Check if Value is Clonable}
\usage{
is.clonable(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
TRUE if the value is clonable, FALSE otherwise.
}
\description{
The \code{is.clonable} function checks if a value is clonable by comparing the "clonable" attribute.
It returns TRUE if the value is clonable, or FALSE otherwise.
}
\examples{
x <- list(clonable = "clonable.0000000WJ3bLybo9e7EvjO8mFmdiHKTX3lh3BYsPiPst")
is.clonable(x) # Returns TRUE

}
