% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.color.class}
\alias{is.color.class}
\title{Check if Value is Color Class}
\usage{
is.color.class(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
TRUE if the value is of the color class, FALSE otherwise.
}
\description{
The \code{is.color.class} function checks if a value is of the color class by comparing the "class" attribute.
It returns TRUE if the value is of the color class, or FALSE otherwise.
}
\examples{
x <- list(class = "color.000000IM9fKDwnkPL8A0dxaV3zT64Q6cjKifaT0BA9kmcvCAQ1NT")
is.color.class(x) # Returns TRUE

}
