% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.group.class}
\alias{is.group.class}
\title{Check if Value is Group Class}
\usage{
is.group.class(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
TRUE if the value is of the group class, FALSE otherwise.
}
\description{
The \code{is.group.class} function checks if a value is of the group class by comparing the "class" attribute.
It returns TRUE if the value is of the group class, or FALSE otherwise.
}
\examples{
x <- list(class = "group.00000004intEgVdJq9HfQwRwDXWRlZYxDSX646IH6s51zkcT1B")
is.group.class(x) # Returns TRUE

}
