% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.nasd}
\alias{is.nasd}
\title{Check if Value is NASD}
\usage{
is.nasd(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
TRUE if the value is the NASD identifier, FALSE otherwise.
}
\description{
The \code{is.nasd} function checks if a value is the NASD identifier and returns TRUE if it is, or FALSE otherwise.
}
\examples{
is.nasd("nasd.00000000n3F8zpfpCoa5NnO6ELu8GhgA86qVNRnGw0gp93ekXeYgaPxu") # Returns TRUE
is.nasd("Hello") # Returns FALSE

}
