% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.null}
\alias{is.null}
\title{Check if Value is Null}
\usage{
is.null(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
TRUE if the value is null or the null identifier, FALSE otherwise.
}
\description{
The \code{is.null} function checks if a value is null or the null identifier and returns TRUE if it is, or FALSE otherwise.
}
\examples{
is.null(NULL) # Returns TRUE
is.null("null.000000000GBxRkEka3IDpOK5h0FNBnzLXMq1AIewoVOn64wincCVyl") # Returns TRUE
is.null("Hello") # Returns FALSE

}
