% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isnt.na}
\alias{isnt.na}
\title{Check if Value is not NA}
\usage{
isnt.na(x, y)
}
\arguments{
\item{x}{The value to check.}

\item{y}{The default value to return if the value is NA.}
}
\value{
The value itself if it is not NA, or the default value if it is NA.
}
\description{
The \code{isnt.na} function checks if a value is not NA and returns the value itself if it is not NA, or a specified default value otherwise.
}
\examples{
isnt.na(5, "Default") # Returns 5
isnt.na(NA, "Default") # Returns "Default"

}
