% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isnt.nasd}
\alias{isnt.nasd}
\title{Check if Value is not NASD}
\usage{
isnt.nasd(x, y)
}
\arguments{
\item{x}{The value to check.}

\item{y}{The default value to return if the value is NASD.}
}
\value{
The value itself if it is not NASD, or the default value if it is NASD.
}
\description{
The \code{isnt.nasd} function checks if a value is not NASD (Not A Significant Data) and returns the value itself if it is not NASD, or a specified default value otherwise.
}
\examples{
isnt.nasd(42, "Default") # Returns 42
isnt.nasd(nasd(), "Default") # Returns "Default"

}
