% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{isnt.null}
\alias{isnt.null}
\title{Check if Value is not Null}
\usage{
isnt.null(x, y)
}
\arguments{
\item{x}{The value to check.}

\item{y}{The default value to return if the value is null.}
}
\value{
The value itself if it is not null, or the default value if it is null.
}
\description{
The \code{isnt.null} function checks if a value is not null and returns the value itself if it is not null, or a specified default value otherwise.
}
\examples{
isnt.null("Hello", "Default") # Returns "Hello"
isnt.null(NULL, "Default") # Returns "Default"

}
