% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{iterVector}
\alias{iterVector}
\title{Create an iterator for a vector}
\usage{
iterVector(vector, cyclic = FALSE, defVal = NA, decreasing = F)
}
\arguments{
\item{vector}{The vector to iterate over.}

\item{cyclic}{Logical indicating whether the iteration should be cyclic, i.e.,
whether it should wrap around to the beginning of the vector after reaching the end.
Default is FALSE.}

\item{defVal}{The default value to return when there are no more elements in the
vector. Default is NA.}

\item{decreasing}{Logical indicating whether the iteration should be in decreasing
order. If TRUE, the iterator starts from the last element and moves towards the
first element. Default is FALSE, indicating increasing order.}
}
\value{
An iterator object for the vector.
}
\description{
This function creates an iterator object that allows iterating over the elements
of a vector. The iterator provides methods for accessing the next element (\code{Next}),
the previous element (\code{Prev}), the current element (\code{This}), checking if there
are more elements (\code{HasNext}), and restarting the iteration (\code{Restart}).
}
\examples{
# Create an iterator for a vector
vector <- c(1, 2, 3, 4, 5)
iter <- iterVector(vector)

# Iterate over the vector
iter$HasNext() # Returns TRUE
iter$Next() # Returns 1
iter$Next() # Returns 2
iter$Next() # Returns 3
iter$Next() # Returns 4
iter$Next() # Returns 5
iter$Next() # Returns NA (end of vector)
iter$HasNext() # Returns FALSE

}
