% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{join}
\alias{join}
\title{Join Strings}
\usage{
join(..., sep = "")
}
\arguments{
\item{...}{One or more character vectors to be concatenated.}

\item{sep}{The separator to be used between the elements. Default is an empty string.}
}
\value{
A single character vector obtained by concatenating the input vectors with the specified separator.
}
\description{
The \code{join} function concatenates multiple strings into a single string using a specified separator.
}
\examples{
join("Hello", "world", sep = ", ") # Returns "Hello, world"
join("a", "b", "c", sep = "-") # Returns "a-b-c"

}
