% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{l}
\alias{l}
\title{Create a list with specified elements}
\usage{
l(...)
}
\arguments{
\item{...}{The elements to be included in the list.}
}
\value{
A list object containing the specified elements.
}
\description{
The \code{l} function creates a list with the specified elements. It is a shorthand way to create a list object in R.
}
\examples{
l(1, 2, 3)
l("a", "b", "c")
l(1, "a", TRUE)

}
