% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Legend}
\alias{Legend}
\title{Legend Function}
\usage{
Legend(name, ...)
}
\arguments{
\item{name}{(character) The name of the parameter.}

\item{...}{Additional arguments to be passed.}
}
\value{
(NULL)
}
\description{
Description:
This function creates a new global parameter for controlling legend properties.
}
\examples{
Legend("mylegend", cex.legend = 1.3, col.legend = "#000000", alp.legend = 1,
       lin = TRUE, col.lin = NULL, alp.lin = 1, lwd.lin = 1, lty.lin = 1,
       col.box = "#BEBEBE", alp.box = 1, lwd.box = 1, lty.box = 1,
       col.back = "#E6E6E6", alp.back = 0.8)

}
\references{
For more information, refer to the documentation of the \code{newGlobalParam} function.
}
\seealso{
\code{\link{newGlobalParam}}
}
\keyword{global}
\keyword{legend}
\keyword{parameter}
