% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Lines}
\alias{Lines}
\title{Lines Function}
\usage{
Lines(name, ...)
}
\arguments{
\item{name}{(character) The name of the parameter.}

\item{...}{Additional arguments to be passed.}
}
\value{
(NULL)
}
\description{
Description:
This function creates a new global parameter for controlling line properties in a plot.
}
\examples{
Lines("mylines", alp.lines = 0.5, col.lines = "blue", lty.lines = 2, lwd.lines = 2)

}
\references{
For more information, refer to the documentation of the \code{newGlobalParam} function.
}
\seealso{
\code{\link{newGlobalParam}}
}
\keyword{global}
\keyword{line}
\keyword{parameter}
\keyword{plot}
