% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{linesGroup}
\alias{linesGroup}
\title{linesGroup: Grouped Lines Plot}
\usage{
linesGroup(
  ...,
  plot.lines = F,
  plot.points = F,
  plot.shadow = F,
  plot.mean = F,
  col = NULL
)
}
\arguments{
\item{...}{The elements to be processed as a list.}

\item{plot.lines}{Logical vector indicating whether to plot lines for each group.
If a single logical value is provided, it will be recycled for each group.}

\item{plot.points}{Logical vector indicating whether to plot points for each group.
If a single logical value is provided, it will be recycled for each group.}

\item{plot.shadow}{Logical vector indicating whether to plot shadows for each group.
If a single logical value is provided, it will be recycled for each group.}

\item{plot.mean}{Logical vector indicating whether to plot means for each group.
If a single logical value is provided, it will be recycled for each group.}

\item{col}{Vector of colors for the groups. If not provided, default colors will be used.}
}
\description{
This function plots grouped lines based on the input data. It allows for customizing the appearance of the lines, points, shadows, and means within each group.
}
\examples{
# Create data for three groups
group1 <- data.frame(x = 1:10, y = 1:10)
group2 <- data.frame(x = 1:10, y = seq(1, 10, by = 2))
group3 <- data.frame(x = 1:10, y = seq(1, 10, by = 3))

# Plot grouped lines with shadows and means
linesGroup(group1, group2, group3, plot.lines = TRUE, plot.shadow = TRUE, plot.mean = TRUE)

}
