% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list.compareItem}
\alias{list.compareItem}
\title{Compare Item in List}
\usage{
list.compareItem(lista, name, item)
}
\arguments{
\item{lista}{The list to search for the item.}

\item{name}{The name of the item in the list.}

\item{item}{The value to compare with the item.}
}
\value{
\code{TRUE} if the item in the list is equal to the specified value, \code{FALSE} otherwise.
}
\description{
The \code{list.compareItem} function compares an item in a list with a specified value and returns \code{TRUE} if they are equal.
}
\examples{
my_list <- list(a = 1, b = 2, c = 3)
list.compareItem(my_list, "a", 1) # Returns TRUE
list.compareItem(my_list, "b", 4) # Returns FALSE

}
