% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{listToVec}
\alias{listToVec}
\title{Convert List Elements to a Vector}
\usage{
listToVec(List, index)
}
\arguments{
\item{List}{The list from which elements will be extracted.}

\item{index}{A numeric vector specifying the indices of the elements to be extracted.}
}
\value{
A vector containing the elements extracted from the list.
}
\description{
The \code{listToVec} function is used to convert specific elements of a list to a vector. It takes a list and an index of the elements to be extracted from the list, and returns a vector containing those elements.
}
\examples{
myList <- list("apple", "banana", "orange", "grape")
listToVec(myList, c(2, 4)) # Extracts the second and fourth elements from the list

}
