% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mabline}
\alias{mabline}
\title{mabline}
\usage{
mabline(
  h = NULL,
  v = NULL,
  col.abline = nasd(),
  alp.abline = nasd(),
  lty.abline = nasd(),
  lwd.abline = nasd(),
  sld.abline = nasd(),
  ...
)
}
\arguments{
\item{h}{the \code{y}-values for horizontal lines.}

\item{v}{the \code{x}-values for vertical lines.}

\item{col.lines}{lines color.}

\item{alp.lines}{lines color transparency.}

\item{lty.lines}{lines type.}

\item{lwd.lines}{lines width.}

\item{sld.lines}{lines density.}
}
\description{
This is a modification to the \code{graphics::abline} function.
}
\examples{
mplot()
mabline(h=0, v=c(-.5,.5), col.abline="red")

mplot()
mabline(h=(-1):1, v=c(-.5,.5), col.abline=c("blue","red"), lwd.abline=5)

mplot()
mabline(h=.5, col.abline=color(1,0,1), lwd.abline=6)
mabline(h=-.5, col.abline=color(1,0,1), lwd.abline=6, sld=F)

sq<- seq(-1,1,.1)
mplot()
mabline(h=sq, v=sq, col.abline=color(1:0,0:1,1))

mplot()
mabline(h=(-1):1, v=c(-.5,.5), col.abline=c("blue","red"), lwd.abline=5)

mplot()
mabline(h=.5, col.abline=color(1,0,1), lwd.abline=6)
mabline(h=-.5, col.abline=color(1,0,1), lwd.abline=6, sld=F)

sq<- seq(-1,1,.1)
mplot()
mabline(h=sq, v=sq, col.abline=color(1:0,0:1,1), lty.abline=3)
}
\seealso{
\code{itz::Abline}, \code{itz::mplot} & \code{itz::color}.
}
