% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mbar}
\alias{mbar}
\title{mbar}
\usage{
mbar(
  x,
  y,
  col.bar = nasd(),
  alp.bar = nasd(),
  lwd.bar = nasd(),
  lty.bar = nasd(),
  horizontal = nasd(),
  ...
)
}
\arguments{
\item{x}{\code{x}-coordinate.}

\item{y}{bar values.}

\item{col.bar}{bar color.}

\item{alp.bar}{bar color transparency.}

\item{lwd.bar}{bar width.}

\item{lty.bar}{bar type.}

\item{horizontal}{horizontal bar.}
}
\description{
This generates a bar plot.
}
\examples{
x<--10:10
y0<-x^2
y1<-2*x+20

mplot(xlim=c(-10,10), ylim=c(0,100))

mlines(x, y0)

mlines(x, list(y0, y1), col.lines="#AA33AA", lty.lines=2)

mlines(x, list(y0, y1), col.lines=c("red","blue"),lwd.lines=3)

mlines(x, y0+30, col=color(0.3,0.7,0.5), lty=2, lwd=3)

mlines(x, list(y0, y1), col=c("red","blue"), lwd=5)
}
\seealso{
\code{itz::Bar} & \code{itz::color}.
}
