% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mlegend}
\alias{mlegend}
\title{mlegend}
\usage{
mlegend(
  pos = "center",
  legend = "",
  cex.legend = nasd(),
  col.legend = nasd(),
  alp.legend = nasd(),
  lin = nasd(),
  col.lin = nasd(),
  alp.lin = nasd(),
  lwd.lin = nasd(),
  lty.lin = nasd(),
  col.box = nasd(),
  alp.box = nasd(),
  lwd.box = nasd(),
  lty.box = nasd(),
  col.back = nasd(),
  alp.back = nasd(),
  ...
)
}
\arguments{
\item{pos}{location of legend.}

\item{legend}{text of legend.}

\item{cex.legend}{scale for legend.}

\item{col.legend}{legend color.}

\item{alp.legend}{legend color transparency.}

\item{lin}{to draw a line or a point.}

\item{col.lin}{lines color.}

\item{alp.lin}{lines color transparency.}

\item{lwd.lin}{lines width.}

\item{lty.lin}{lines type.}

\item{col.box}{box color.}

\item{alp.box}{box color transparency.}

\item{lwd.box}{box width.}

\item{lty.box}{box type.}

\item{col.back}{back color.}

\item{alp.back}{back color transparency.}
}
\description{
This is a modification to the \code{graphics::legend} function.
}
\examples{
mplot()
mlegend(legend=c("a","b"), col.lin=c("blue","red"), lty.lin=c(1,2))

mlegend("topleft", legend=c("a","b","c"), col.lin=c("blue","red","green"), lty.lin=c(3,2,1), lwd.lin=c(1,2,3))

mlegend("topright", legend=c("a","b"), col.lin=c("purple","magenta"), lwd.lin=4, lin=c(F,T))

mplot()
mlegend(c(0,1),legend=c("a","b"), col.lin=c("blue","red"), cex.legend=2, col.box="orange", lwd.box=5, lty.box=1)
}
\seealso{
\code{itz::Legend}, \code{itz::mplot} & \code{itz::color}.
}
