% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mlines}
\alias{mlines}
\title{mlines}
\usage{
mlines(
  x = NULL,
  y = NULL,
  col.lines = nasd(),
  alp.lines = nasd(),
  lty.lines = nasd(),
  lwd.lines = nasd(),
  ...
)
}
\arguments{
\item{x}{\code{x}-coordinate.}

\item{y}{\code{y}-coordinate.}

\item{col.lines}{lines color.}

\item{alp.lines}{lines color transparency.}

\item{lty.lines}{lines type.}

\item{lwd.lines}{lines width.}
}
\description{
This is a modification to the \code{graphics::lines} function.
}
\details{
\code{x} and \code{y} could be vectors or lists.
}
\examples{
x<--10:10
y0<-x^2
y1<-2*x+20

mplot(xlim=c(-10,10), ylim=c(0,100))

mlines(x, y0)

mlines(x, list(y0, y1), col.lines="#AA33AA", lty.lines=2)

mlines(x, list(y0, y1), col.lines=c("red","blue"),lwd.lines=3)

mlines(x, y0+30, col=color(0.3,0.7,0.5), lty=2, lwd=3)

mlines(x, list(y0, y1), col=c("red","blue"), lwd=5)
}
\seealso{
\code{itz::Lines}, \code{itz::mplot} & \code{itz::color}.
}
