% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mplot}
\alias{mplot}
\title{mplot}
\usage{
mplot(
  x = NULL,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  main = NA,
  xlab = NA,
  ylab = NA,
  alp.background = nasd(),
  alp.backlines = nasd(),
  alp.border = nasd(),
  cex.axis = nasd(),
  cex.lab = nasd(),
  cex.main = nasd(),
  col.background = nasd(),
  col.backlines = nasd(),
  col.border = nasd(),
  labX = nasd(),
  labY = nasd(),
  lty.backlines = nasd(),
  lwd.backlines = nasd(),
  lx = nasd(),
  ly = nasd(),
  plot.lines = nasd(),
  plot.points = nasd(),
  px = nasd(),
  py = nasd(),
  xdigits = nasd(),
  ydigits = nasd(),
  xaxis = nasd(),
  yaxis = nasd(),
  xaxis.pos = nasd(),
  yaxis.pos = nasd(),
  ...
)
}
\arguments{
\item{x}{\code{x}-coordinate.}

\item{y}{\code{y}-coordinate.}

\item{xlim}{limits for \code{x}-axis.}

\item{ylim}{limits for \code{x}-axis.}

\item{col}{color for \code{itz::mlines} and \code{itz::mpoints}.}

\item{main}{main title.}

\item{xlab}{\code{x}-axis legend.}

\item{ylab}{\code{y}-axis legend.}

\item{alp.background}{background color transparency.}

\item{alp.backlines}{backlines color transparency.}

\item{alp.border}{border color transparency.}

\item{cex.axis}{scale for \code{y}-axis legend.}

\item{cex.lab}{scale for \code{x}-axis legend.}

\item{cex.main}{scale for main title.}

\item{col.background}{background color.}

\item{col.backlines}{backlines color.}

\item{col.border}{border color.}

\item{labX}{legends of \code{x}-axis backlines.}

\item{labY}{legends of \code{y}-axis backlines.}

\item{lty.backlines}{backlines type.}

\item{lwd.backlines}{backlines width.}

\item{lx}{values of \code{x}-axis backlines.}

\item{ly}{values of \code{y}-axis backlines.}

\item{plot.lines}{to plot lines graphic.}

\item{plot.points}{to plot points graphic.}

\item{px}{number of \code{x}-axis backlines.}

\item{py}{number of \code{y}-axis backlines.}

\item{xdigits}{\code{x}-axis rounding digits.}

\item{ydigits}{\code{y}-axis rounding digits.}

\item{xaxis}{to show the \code{x}-axis.}

\item{yaxis}{to show the \code{y}-axis.}

\item{xaxis.pos}{to show the \code{x}-axis on bottom side.}

\item{yaxis.pos}{to show the \code{y}-axis on left side.}

\item{...}{supports \code{mlines} and \code{mpoints} parameters}
}
\description{
This is a modification to the \code{graphics::plot} function.
}
\examples{
mplot()

mplot(0:1)

sq<-seq(0, 4*pi, 0.6)
mplot(sq, sin(sq), plot.points=T, ydigits=2, col="magenta", col.points="cyan")

t<-seq(0, 1, 0.05)
mplot(t, l(sqrt(t),t, t^2), plot.points=T, plot.lines=F, col=c("red","blue","green"), lwd.points=2)

mplot(rnorm(100), px=3, py=4, ylim=c(-2.5,2.5), ydigits=0, xdigits=0)

mplot(runif(100),rnorm(100), px=2, py=2, xaxis=F, yaxis=F, col.background=color(.21,.05,.2), col.backlines="#C9F2CC",xlab="runif", ylab="rnorm", main="title", lwd.backlines=4)
}
\seealso{
\code{itz::Plot}, \code{itz::mlines}, \code{itz::mpoints} & \code{itz::color}.
}
