% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mpoints}
\alias{mpoints}
\title{mpoints}
\usage{
mpoints(
  x = NULL,
  y = NULL,
  alp.points = nasd(),
  col.points = nasd(),
  lwd.points = nasd(),
  ...
)
}
\arguments{
\item{x}{\code{x}-coordinate.}

\item{y}{\code{y}-coordinate.}

\item{alp.points}{points color transparency.}

\item{col.points}{points color.}

\item{lwd.points}{points width.}
}
\description{
This is a modification to the \code{graphics::points} function.
}
\details{
\code{x} and \code{y} could be vectors or lists.
}
\examples{
x<--10:10
y0<-x^2
y1<-2*x+20

mplot(xlim=c(-10,10), ylim=c(0,100))

mpoints(x, y0)

mpoints(x, list(y0, y1), col.points="#AA33AA")

mpoints(x, list(y0, y1), col.points=c("red","blue"),lwd.points=3)

mpoints(x, y0+30, col=color(0.3,0.7,0.5), lwd=3)

mpoints(x, list(y0, y1), col=c("red","blue"), lwd=5)
}
\seealso{
\code{itz::Points}, \code{itz::mplot} & \code{itz::color}.
}
