% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mseq}
\alias{mseq}
\title{Generate a modified sequence}
\usage{
mseq(from, to, n = 2, close = T, by = NULL, ...)
}
\arguments{
\item{from}{The starting point of the sequence.}

\item{to}{The ending point of the sequence.}

\item{n}{The number of elements in the sequence. Default is 2.}

\item{close}{A logical value indicating whether the sequence should include the endpoint (\code{to}) or not. Default is \code{TRUE}.}

\item{by}{The step size between consecutive elements in the sequence. If \code{NULL}, the step size is automatically determined based on the number of elements (\code{n}) and the range between \code{from} and \code{to}. Default is \code{NULL}.}

\item{...}{Additional arguments to be passed to the \code{seq} function.}
}
\value{
A sequence of numbers between \code{from} and \code{to}, with the specified number of elements (\code{n}), step size (\code{by}), and endpoint inclusion (\code{close}).
}
\description{
The \code{mseq} function generates a modified sequence of numbers between a starting point and an ending point. It provides flexibility in defining the number of elements in the sequence and the step size.
}
\examples{
mseq(1, 10, n = 5)
mseq(1, 10, n = 5, close = FALSE)
mseq(1, 10, by = 2)

}
