% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nasd}
\alias{nasd}
\title{NASD (Not A Significant Data) Identifier}
\usage{
nasd()
}
\value{
The NASD identifier.
}
\description{
The \code{nasd} function returns the NASD (Not A Significant Data) identifier.
}
\examples{
nasd() # Returns "nasd.00000000n3F8zpfpCoa5NnO6ELu8GhgA86qVNRnGw0gp93ekXeYgaPxu"

}
