% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{newFunction}
\alias{newFunction}
\title{Create a new function object}
\usage{
newFunction(...)
}
\arguments{
\item{...}{Arguments to be used as the function's arguments.}
}
\value{
A list containing methods to manipulate and execute the generated function.
}
\description{
This function creates a new function object by generating a random file name and
defining a function code based on the provided arguments.
}
\examples{
# Create a new function object
f <- newFunction(x, y, z)

# Add code to the function
f$add("result <- x + y * z")

# Get the generated code
f$getCode()
# Returns the code:
# "function(x, y, z) {
#    result <- x + y * z
# }"

# Compile and execute the function
result <- f$run(2, 3, 4)
# Returns the result: 14

}
