% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{newList}
\alias{newList}
\title{Create a new list object}
\usage{
newList(...)
}
\arguments{
\item{...}{Optional elements to initialize the list object.}
}
\value{
A new instance of the \code{.vector} class wrapped in a list.
}
\description{
This function creates a new instance of the \code{.vector} class and wraps it in a list.
It can optionally be initialized with one or more elements.
}
\examples{
# Create an empty list object
l1 <- newList()
l1$length() # Returns the length of the list: 0

# Create a list object with initial elements
l2 <- newList("a", 1, TRUE)
l2$get() # Returns the list ["a", 1, TRUE]
l2$length() # Returns the length of the list: 3

# Get values from the list
l2$get(2) # Returns the element at index 2: 1
l2$get("a") # Returns the element with name "a": NULL
l2$get(4) # Returns NULL since the list has no element at index 4

}
