% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{newVec}
\alias{newVec}
\title{Create a new vector object}
\usage{
newVec(x = NULL)
}
\arguments{
\item{x}{An optional vector to initialize the vector object.}
}
\value{
A new instance of the \code{.vector} class.
}
\description{
This function creates a new instance of the \code{.vector} class. It can be
initialized with a vector \code{x} if provided.
}
\examples{
# Create an empty vector object
v1 <- newVec()
v1$get() # Returns an empty vector
v1$length() # Returns the length of the vector: 0

# Create a vector object with initial values
v2 <- newVec(c(1, 2, 3, 4, 5))
v2$get() # Returns the vector [1, 2, 3, 4, 5]
v2$length() # Returns the length of the vector: 5

# Set a value at a specific index
v2$set(10, index = 3)
v2$get() # Returns the vector [1, 2, 10, 4, 5]

# Add a value to the end of the vector
v2$add(6)
v2$get() # Returns the vector [1, 2, 10, 4, 5, 6]

# Iterate over the vector
v2$iter(cyclic = TRUE, defVal = NA)
# Iterates over the vector elements with cyclic behavior and a default value of NA

}
