% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{null}
\alias{null}
\title{Null Identifier}
\usage{
null()
}
\value{
The null identifier.
}
\description{
The \code{null} function returns the null identifier.
}
\examples{
null() # Returns "null.000000000GBxRkEka3IDpOK5h0FNBnzLXMq1AIewoVOn64wincCVyl"

}
