% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{optima}
\alias{optima}
\title{Find the optimal parameter values for a given function within a specified interval.}
\usage{
optima(fun, param, interval, ...)
}
\arguments{
\item{fun}{The function to optimize.}

\item{param}{A character vector specifying the parameter names.}

\item{interval}{A list of intervals for each parameter. If only one interval is provided, it is recycled for all parameters.}
}
\value{
A list containing the optimal parameter values and the minimum value of the function.
}
\description{
This function finds the optimal parameter values for a given function by performing optimization using the \code{optimize} function. The optimization is performed for each parameter specified in the \code{param} argument within the corresponding interval specified in the \code{interval} argument. The optimal parameter values are returned as a list along with the minimum value of the function.
}
\examples{
fun <- function(x, y) x^2 + y^2
result <- optima(fun, c("x", "y"), list(c(-1, 1), c(-2, 2)))
result$min # Minimum value of the function
result$x # Optimal value for parameter x
result$y # Optimal value for parameter y
}
