% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{optimizer}
\alias{optimizer}
\title{Find the optimal parameter values for a given function within a specified interval using an adaptive optimization algorithm.}
\usage{
optimizer(
  f,
  interval,
  names = NULL,
  min.partitions = 5,
  max.partitions = min.partitions + length(interval),
  ...
)
}
\arguments{
\item{f}{The function to optimize.}

\item{interval}{A list of intervals for each parameter.}

\item{names}{A character vector specifying the names of the parameters.}

\item{min.partitions}{The minimum number of partitions for adaptive optimization.}

\item{max.partitions}{The maximum number of partitions for adaptive optimization.}

\item{...}{Additional arguments passed to the function \code{f}.}
}
\value{
A list containing information about the optimization process, including the number of iterations, precision, and the optimal parameter values.
}
\description{
This function finds the optimal parameter values for a given function using an adaptive optimization algorithm. It performs optimization for each parameter specified in the \code{names} argument within the corresponding interval specified in the \code{interval} argument. The optimization algorithm adjusts the precision of the intervals based on the variation in the function values to achieve accurate results. The optimal parameter values, along with other information such as the number of iterations and precision, are returned as a list.
}
\examples{
fun <- function(x, y) x^2 + y^2
result <- optimizer(fun, list(c(-1, 1), c(-2, 2)), c("x", "y"))
result$loops # Number of iterations
result$fmin.prescision # Precision of the minimum value
result$var.prescision # Precision of each parameter
result$optim # Optimal parameter values and minimum value
result$interval # Final interval for each parameter
}
