% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{Plot Function}
\usage{
Plot(name, ...)
}
\arguments{
\item{name}{(character) The name of the parameter.}

\item{...}{Additional arguments to be passed.}
}
\value{
(NULL)
}
\description{
Description:
This function creates a new global parameter for controlling plot properties.
}
\examples{
Plot("myplot", alp.backlines = 0.5, alp.background = 1, alp.border = 0, cex.axis = 1.1,
     cex.lab = 1.3, cex.main = 1.3, col.background = "#E0E0E0", col.backlines = "white",
     col.border = "black", labX = NA, labY = NA, lty.backlines = 1, lwd.backlines = 1,
     lx = NA, ly = NA, plot.lines = TRUE, plot.points = FALSE, px = 6, py = 6,
     xdigits = NA, ydigits = NA, xaxis = TRUE, yaxis = TRUE, xaxis.pos = TRUE, yaxis.pos = TRUE)

}
\references{
For more information, refer to the documentation of the \code{newGlobalParam} function.
}
\seealso{
\code{\link{newGlobalParam}}
}
\keyword{global}
\keyword{parameter}
\keyword{plot}
