% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Points}
\alias{Points}
\title{Points Function}
\usage{
Points(name, ...)
}
\arguments{
\item{name}{(character) The name of the parameter.}

\item{...}{Additional arguments to be passed.}
}
\value{
(NULL)
}
\description{
Description:
This function creates a new global parameter for controlling point properties in a plot.
}
\examples{
Points("mypoints", alp.points = 0.5, col.points = "red", lwd.points = 2)

}
\references{
For more information, refer to the documentation of the \code{newGlobalParam} function.
}
\seealso{
\code{\link{newGlobalParam}}
}
\keyword{global}
\keyword{parameter}
\keyword{plot}
\keyword{point}
