% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{printer}
\alias{printer}
\title{printer: PDF Printer}
\usage{
printer(dir = "", prop = 0.68, width = 9, height = NULL, size = c(1, 1))
}
\arguments{
\item{dir}{Directory where the PDF files will be saved. If not provided, the files will be saved in the current working directory.}

\item{prop}{Proportion of the width to the height of the PDF page.}

\item{width}{Width of the PDF page in inches.}

\item{height}{Height of the PDF page in inches.}

\item{size}{Vector specifying the number of rows and columns for arranging multiple plots on a single page.
The first element represents the number of rows, and the second element represents the number of columns.}
}
\value{
A list of functions that can be used to control the PDF printer.
}
\description{
This function creates a PDF printer that allows you to easily generate PDF files with custom settings.
}
\details{
The PDF printer provides the following functions:
\itemize{
\item \code{PDF}: Generates a PDF file with the specified settings and plots.
\item \code{ON}: Activates the PDF printer, allowing plots to be saved as PDF files.
\item \code{OFF}: Deactivates the PDF printer, allowing plots to be displayed on the screen instead.
}
}
\examples{
# Create a PDF printer with default settings
myprinter <- printer()

# Activate the PDF printer
myprinter$ON()

# Generate a PDF file with a plot
myprinter$PDF(plot(1:10))

# Deactivate the PDF printer
myprinter$OFF()

}
