% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pst}
\alias{pst}
\title{Paste with Separator}
\usage{
pst(..., sep = "")
}
\arguments{
\item{...}{One or more character vectors to be concatenated.}

\item{sep}{The separator to be used between the elements. Default is an empty string.}
}
\value{
A single character vector obtained by concatenating the input vectors with the specified separator.
}
\description{
The \code{pst} function is a wrapper around the base \code{paste} function that allows specifying a custom separator.
}
\examples{
pst("Hello", "world", sep = ", ") # Returns "Hello, world"

}
