% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{randFile}
\alias{randFile}
\title{Generate a Random File Name}
\usage{
randFile()
}
\value{
A character string representing a random file name.
}
\description{
The \code{randFile} function generates a random file name by combining a prefix, a sequence of zeros, and a random selection of characters.
}
\examples{
randFile() # Returns something like "0x00000A1b3FgH"
randFile() # Returns something like "0x00000CB2xYzTg"

}
