% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Shadow}
\alias{Shadow}
\title{Shadow Function}
\usage{
Shadow(name, ...)
}
\arguments{
\item{name}{(character) The name of the parameter.}

\item{...}{Additional arguments to be passed.}
}
\value{
(NULL)
}
\description{
Description:
This function creates a new global parameter for controlling shadow plot properties.
}
\examples{
Shadow("myshadow", alp.border = 1, alp.shadow = 0.5, alp.mean = 1,
       col.border = NULL, col.shadow = NULL, col.mean = NULL,
       lty.border = 1, lty.mean = 1, lwd.border = 0, lwd.mean = 1,
       plot.shadow = TRUE, plot.mean = TRUE)

}
\references{
For more information, refer to the documentation of the \code{newGlobalParam} function.
}
\seealso{
\code{\link{newGlobalParam}}
}
\keyword{global}
\keyword{parameter}
\keyword{shadow}
