% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{simSample}
\alias{simSample}
\title{Generate a simulated sample from a given dataset or empirical cumulative distribution function.}
\usage{
simSample(x, Nsim = NA)
}
\arguments{
\item{x}{The input dataset.}

\item{Nsim}{The size of the simulated sample.}
}
\value{
If \code{Nsim} is provided, a simulated sample of size \code{Nsim}. If \code{Nsim} is not provided, a closure function for generating simulated samples.
}
\description{
This function generates a simulated sample from a given dataset or empirical cumulative distribution function (ecdf). If the argument \code{Nsim} is specified, it returns a simulated sample of size \code{Nsim} using the ecdf of the input dataset \code{x}. If \code{Nsim} is not provided, it returns a closure function that takes \code{Nsim} as an argument and generates a simulated sample of size \code{Nsim} using the stored ecdf.
}
\examples{
data <- c(1, 2, 3, 4, 5)
sample <- simSample(data)
sample(10) # Generate a simulated sample of size 10
sample(100) # Generate a simulated sample of size 100
}
