% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{simTable}
\alias{simTable}
\title{Generate simulated samples from a given table or dataset.}
\usage{
simTable(tab, Nsim = NA)
}
\arguments{
\item{tab}{The input table or dataset.}

\item{Nsim}{The size of the simulated samples.}
}
\value{
If \code{Nsim} is provided, a simulated table where each column contains a simulated sample of size \code{Nsim}. If \code{Nsim} is not provided, a closure function for generating simulated tables.
}
\description{
This function generates simulated samples from a given table or dataset. If the argument \code{Nsim} is specified, it returns a simulated table where each column contains a simulated sample of size \code{Nsim} based on the corresponding column of the input table \code{tab}. If \code{Nsim} is not provided, it returns a closure function that takes \code{Nsim} as an argument and generates a simulated table with samples of size \code{Nsim} based on the stored samples from the input table.
}
\examples{
data <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6))
sim <- simTable(data)
sim(10) # Generate a simulated table with samples of size 10
sim(100) # Generate a simulated table with samples of size 100
}
