% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{subList}
\alias{subList}
\title{Extract Sublist from a List}
\usage{
subList(List, index)
}
\arguments{
\item{List}{The original list from which the sublist will be extracted.}

\item{index}{A numeric vector specifying the indices of the elements to be included in the sublist.}
}
\value{
A new list containing the elements specified by the index.
}
\description{
The \code{subList} function is used to extract a sublist from a given list. It takes a list and an index of the elements to be included in the sublist, and returns a new list containing only those elements.
}
\examples{
myList <- list("apple", "banana", "orange", "grape")
subList(myList, c(2, 4)) # Extracts the second and fourth elements from the list

}
