% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{substrae.LIST.DAT}
\alias{substrae.LIST.DAT}
\title{substrae.LIST.DAT: Subtraction of Elements from List and Creation of New Data Object}
\usage{
substrae.LIST.DAT(..., CONDITION, TRANSFORM)
}
\arguments{
\item{...}{The elements to be processed as a list.}

\item{CONDITION}{The condition to determine which elements to subtract from the list.
It should be a logical expression that evaluates to TRUE or FALSE for each element in the list.
Elements that satisfy the condition will be removed from the list.}

\item{TRANSFORM}{The transformation function to be applied to the selected elements.
It should be a function that takes an element as input and returns a transformed version of the element.
The transformed elements will be stored in a new data object.}
}
\value{
A list containing two elements:
\itemize{
\item "lista.123": The updated list after subtracting the selected elements.
\item "data.123": The new data object containing the transformed elements.
}
}
\description{
This function subtracts elements from a list based on a given condition and creates a new data object by transforming the selected elements.
}
\examples{
# Create a list of numbers
numbers <- list(1, 2, 3, 4, 5)

# Subtract even numbers from the list and create a new data object with the squared values
result <- substrae.LIST.DAT(numbers, CONDITION = function(x) x \%\% 2 == 0, TRANSFORM = function(x) x^2)
result$lista.123  # Updated list after subtracting even numbers
result$data.123   # New data object with squared values

}
\keyword{data}
\keyword{list}
\keyword{subtraction}
\keyword{transformation}
