% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tableLinesInter}
\alias{tableLinesInter}
\title{Calculate interpolated values for multiple line segments.}
\usage{
tableLinesInter(X, Y)
}
\arguments{
\item{X}{A list of x-coordinates for the line segments.}

\item{Y}{A list of y-coordinates for the line segments.}
}
\value{
A list containing the interpolated values and additional summary statistics.
}
\description{
This function calculates interpolated values for multiple line segments defined by X and Y coordinates. Given a list of x-coordinates and y-coordinates, the function calculates the interpolated values at common points along the line segments. It returns a list containing the interpolated values, along with additional summary statistics.
}
\examples{
x1 <- c(0, 1, 2, 3)
y1 <- c(0, 2, 4, 6)
x2 <- c(1, 2, 3, 4)
y2 <- c(2, 4, 6, 8)
interpolated <- tableLinesInter(list(x1, x2), list(y1, y2))
interpolated$z # Common points along the line segments
interpolated$tab # Interpolated values for each line segment at the common points
interpolated$min # Minimum interpolated value at each common point
interpolated$max # Maximum interpolated value at each common point
interpolated$mean # Mean interpolated value at each common point
interpolated$sd # Standard deviation of interpolated values at each common point
}
