% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{test}
\alias{test}
\title{Perform benchmark testing of functions X and Y.}
\usage{
test(n, X, Y, ...)
}
\arguments{
\item{n}{The number of times to repeat the benchmark.}

\item{X}{The function to benchmark.}

\item{Y}{The function to benchmark.}
}
\value{
A list containing two vectors of execution times for X and Y.
}
\description{
This function performs benchmark testing of two functions, X and Y, by measuring the execution time of each function. The execution time is measured using the \code{getSeconds} function. The benchmarking is repeated \code{n} times, and the execution times for X and Y are recorded. The results are returned as a list with two vectors, \code{x} and \code{y}, containing the execution times for X and Y, respectively.
}
\examples{
fun1 <- function(x) sqrt(x)
fun2 <- function(x) log(x)
results <- test(10, fun1, fun2, 100, 1000)
results$x # Execution times for function fun1
results$y # Execution times for function fun2
}
