% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mabline}
\alias{mabline}
\title{mabline: Multiple Abline Plotting Function}
\usage{
mabline(
  h = NULL,
  v = NULL,
  col.abline = nasd(),
  alp.abline = nasd(),
  lty.abline = nasd(),
  lwd.abline = nasd(),
  sld.abline = nasd(),
  ...
)
}
\arguments{
\item{h}{A vector of horizontal line positions.}

\item{v}{A vector of vertical line positions.}

\item{col.abline}{The color(s) of the lines. Default is NA.}

\item{alp.abline}{The transparency level of the lines. Default is NA.}

\item{lty.abline}{The line type(s) of the lines. Default is NA.}

\item{lwd.abline}{The line width(s) of the lines. Default is NA.}

\item{sld.abline}{Logical value indicating whether the lines should be solid
or dashed. Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create multiple horizontal or vertical lines on a plot
using the \code{abline} function. It supports customization of line colors, line transparency,
line types, line widths, and whether the lines should be solid or dashed.
}
\examples{
x <- 1:10
y <- x^2
plot(x, y)
mabline(h = 50, col.abline = "red", lwd.abline = 2)
mabline(v = 5:7, col.abline = "blue", lty.abline = 2)

}
\references{
For more information, see the documentation of the \code{abline} function in the base package.
}
\seealso{
\code{\link{abline}}
}
\keyword{lines}
\keyword{multiple}
\keyword{plotting}
