% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mbar}
\alias{mbar}
\title{mbar: Multiple Bar Plot Function}
\usage{
mbar(
  x,
  y,
  col.bar = nasd(),
  alp.bar = nasd(),
  lwd.bar = nasd(),
  lty.bar = nasd(),
  horizontal = nasd(),
  ...
)
}
\arguments{
\item{x}{A numeric vector representing the x-axis values.}

\item{y}{A numeric vector representing the height of the bars.}

\item{col.bar}{The color of the bars. Default is NULL.}

\item{alp.bar}{The transparency level of the bars. Default is NA.}

\item{lwd.bar}{The line width of the bars. Default is NA.}

\item{lty.bar}{The line type of the bars. Default is NA.}

\item{horizontal}{Logical value indicating whether to plot horizontal bars.
Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create multiple bar plots with customizable settings
such as bar color, transparency, line width, line type, and orientation. It supports
plotting multiple sets of x-y data as bars and allows for horizontal or vertical orientation.
}
\examples{
x <- c("A", "B", "C")
y <- c(10, 15, 12)
mbar(x, y, col.bar = "blue", alp.bar = 0.5, lwd.bar = 2)

}
\seealso{
\code{\link{barplot}}
}
\keyword{barplot}
\keyword{multiple}
\keyword{plotting}
