% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mhist}
\alias{mhist}
\title{mhist}
\usage{
mhist(
  x,
  relative = nasd(),
  horizontal = nasd(),
  new.plot = nasd(),
  alp.hist = nasd(),
  col.hist = nasd(),
  lty.hist = nasd(),
  lwd.hist = nasd(),
  ...
)
}
\arguments{
\item{x}{\code{x}-coordinate.}

\item{relative}{relative values.}

\item{horizontal}{horizontal histogram.}

\item{alp.hist}{histogram color transparency.}

\item{col.hist}{histogram color.}

\item{lty.hist}{histogram type.}

\item{lwd.hist}{histogram width.}

\item{new.histogram}{to generate a new plot.}
}
\description{
This is a modification to the \code{graphics::hist} function.
}
\examples{
#
}
\seealso{
\code{itz::Hist}, \code{itz::mplot} & \code{itz::color}.
}
