% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mhist}
\alias{mhist}
\title{mhist: Multiple Histogram Plot Function}
\usage{
mhist(
  x,
  relative = nasd(),
  horizontal = nasd(),
  new.plot = nasd(),
  alp.hist = nasd(),
  col.hist = nasd(),
  lty.hist = nasd(),
  lwd.hist = nasd(),
  ...
)
}
\arguments{
\item{x}{A numeric vector representing the data values for which the histogram will be created.}

\item{relative}{Logical value indicating whether to plot relative frequencies instead of counts.
Default is NA.}

\item{horizontal}{Logical value indicating whether to create a horizontal histogram.
Default is NA.}

\item{new.plot}{Logical value indicating whether to create a new plot. If TRUE, a new plot will be created,
and any existing plot settings (e.g., xlim, ylim) will be ignored. Default is NA.}

\item{alp.hist}{The transparency level of the histogram. Default is NA.}

\item{col.hist}{The color of the histogram. Default is NA.}

\item{lty.hist}{The line type of the histogram. Default is NA.}

\item{lwd.hist}{The line width of the histogram. Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create multiple histogram plots with customizable settings
such as relative frequencies, orientation (horizontal or vertical), new plot creation,
histogram transparency, histogram color, line type, line width, and additional parameters.
It supports plotting histograms for a single variable.
}
\examples{
x <- rnorm(100)
mhist(x, col.hist = "skyblue", alp.hist = 0.7, lwd.hist = 2)

}
\seealso{
\code{\link{hist}}, \code{\link{mbar}}, \code{\link{mplot}}
}
\keyword{histogramplot}
\keyword{multiple}
\keyword{plotting}
