% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mlegend}
\alias{mlegend}
\title{mlegend: Multiple Legend Plot Function}
\usage{
mlegend(
  pos = "center",
  legend = "",
  cex.legend = nasd(),
  col.legend = nasd(),
  alp.legend = nasd(),
  lin = nasd(),
  col.lin = nasd(),
  alp.lin = nasd(),
  lwd.lin = nasd(),
  lty.lin = nasd(),
  col.box = nasd(),
  alp.box = nasd(),
  lwd.box = nasd(),
  lty.box = nasd(),
  col.back = nasd(),
  alp.back = nasd(),
  ...
)
}
\arguments{
\item{pos}{The position of the legend. Can be a character string indicating a predefined position
("topleft", "topright", "bottomleft", "bottomright", "left", "right", "bottom", "top", "center"),
or a numeric vector specifying the x and y coordinates of the legend. Default is "center".}

\item{legend}{The text labels for the legend.}

\item{cex.legend}{The size of the legend text. Default is NA.}

\item{col.legend}{The color of the legend text. Default is NA.}

\item{alp.legend}{The transparency level of the legend text. Default is NA.}

\item{lin}{The line types for the legend lines. Default is NA.}

\item{col.lin}{The colors of the legend lines. Default is NA.}

\item{alp.lin}{The transparency levels of the legend lines. Default is NA.}

\item{lwd.lin}{The line widths of the legend lines. Default is NA.}

\item{lty.lin}{The line types of the legend lines. Default is NA.}

\item{col.box}{The color of the legend box. Default is NA.}

\item{alp.box}{The transparency level of the legend box. Default is NA.}

\item{lwd.box}{The line width of the legend box. Default is NA.}

\item{lty.box}{The line type of the legend box. Default is NA.}

\item{col.back}{The background color of the legend. Default is NA.}

\item{alp.back}{The transparency level of the legend background. Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create multiple legends in a plot with customizable settings
such as legend position, text size, text color, line type, line width, box color, box transparency,
background color, and additional parameters.
}
\examples{
mlegend(pos = "topright", legend = c("Group 1", "Group 2", "Group 3"),
        col.legend = c("red", "blue", "green"), lwd.lin = 2, col.box = "gray")

}
\seealso{
\code{\link{legend}}, \code{\link{posLegend}}
}
\keyword{legendplot}
\keyword{multiple}
\keyword{plotting}
