% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mlines}
\alias{mlines}
\title{mlines: Multiple Lines Plotting Function}
\usage{
mlines(
  x = NULL,
  y = NULL,
  col.lines = nasd(),
  alp.lines = nasd(),
  lty.lines = nasd(),
  lwd.lines = nasd(),
  ...
)
}
\arguments{
\item{x}{A vector or list of x-coordinates.}

\item{y}{A vector or list of y-coordinates.}

\item{col.lines}{The color(s) of the lines. Default is NA.}

\item{alp.lines}{The transparency level of the lines. Default is NA.}

\item{lty.lines}{The line type(s) of the lines. Default is NA.}

\item{lwd.lines}{The line width(s) of the lines. Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create a multiple lines plot by specifying the
x and y coordinates. It supports customization of line colors, line
transparency, line types, and line widths.
}
\examples{
x <- 1:10
y1 <- x
y2 <- 2*x
y3 <- 3*x
mlines(x, y1, col.lines = "red", alp.lines = 0.5)
mlines(x, y2, col.lines = "blue", lty.lines = 2)
mlines(x, y3, col.lines = "green", lwd.lines = 2)

}
\references{
For more information, see the documentation of the \code{lines} function in the base package.
}
\seealso{
\code{\link{lines}}
}
\keyword{lines}
\keyword{multiple}
\keyword{plotting}
