% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mplot}
\alias{mplot}
\title{mplot: Multiple Plot Function}
\usage{
mplot(
  x = NULL,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  main = NA,
  xlab = NA,
  ylab = NA,
  alp.background = nasd(),
  alp.backlines = nasd(),
  alp.border = nasd(),
  cex.axis = nasd(),
  cex.lab = nasd(),
  cex.main = nasd(),
  col.background = nasd(),
  col.backlines = nasd(),
  col.border = nasd(),
  labX = nasd(),
  labY = nasd(),
  lty.backlines = nasd(),
  lwd.backlines = nasd(),
  lx = nasd(),
  ly = nasd(),
  plot.lines = nasd(),
  plot.points = nasd(),
  px = nasd(),
  py = nasd(),
  xdigits = nasd(),
  ydigits = nasd(),
  xaxis = nasd(),
  yaxis = nasd(),
  xaxis.pos = nasd(),
  yaxis.pos = nasd(),
  ...
)
}
\arguments{
\item{x}{A list of x values for each data set. If a single vector is provided,
it will be treated as a single data set.}

\item{y}{A list of y values for each data set. If a single vector is provided,
it will be treated as a single data set.}

\item{xlim}{The x-axis limits of the plot. Default is NULL.}

\item{ylim}{The y-axis limits of the plot. Default is NULL.}

\item{col}{The color(s) to use for the plot. Default is NULL.}

\item{main}{The main title of the plot. Default is NA.}

\item{xlab}{The label for the x-axis. Default is NA.}

\item{ylab}{The label for the y-axis. Default is NA.}

\item{alp.background}{The transparency level of the plot background. Default is NA.}

\item{alp.backlines}{The transparency level of the background grid lines. Default is NA.}

\item{alp.border}{The transparency level of the plot border. Default is NA.}

\item{cex.axis}{The font size for the axis labels. Default is NA.}

\item{cex.lab}{The font size for the axis labels. Default is NA.}

\item{cex.main}{The font size for the plot title. Default is NA.}

\item{col.background}{The color of the plot background. Default is NA.}

\item{col.backlines}{The color of the background grid lines. Default is NA.}

\item{col.border}{The color of the plot border. Default is NA.}

\item{labX}{The label for the x-axis tick marks. Default is NA.}

\item{labY}{The label for the y-axis tick marks. Default is NA.}

\item{lty.backlines}{The line type of the background grid lines. Default is NA.}

\item{lwd.backlines}{The line width of the background grid lines. Default is NA.}

\item{lx}{The position of the x-axis tick marks. Default is NA.}

\item{ly}{The position of the y-axis tick marks. Default is NA.}

\item{plot.lines}{Logical value indicating whether to plot lines. Default is NA.}

\item{plot.points}{Logical value indicating whether to plot points. Default is NA.}

\item{px}{The number of x-axis tick marks. Default is NA.}

\item{py}{The number of y-axis tick marks. Default is NA.}

\item{xdigits}{The number of digits to display on the x-axis tick marks. Default is NA.}

\item{ydigits}{The number of digits to display on the y-axis tick marks. Default is NA.}

\item{xaxis}{Logical value indicating whether to display the x-axis. Default is NA.}

\item{yaxis}{Logical value indicating whether to display the y-axis. Default is NA.}

\item{xaxis.pos}{Logical value indicating whether to display the x-axis tick marks.
Default is NA.}

\item{yaxis.pos}{Logical value indicating whether to display the y-axis tick marks.
Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create multiple plots with customizable settings
such as axis labels, plot title, background color, line and point properties,
and more. It supports plotting multiple sets of x-y data, specifying the x and
y limits, and customizing various visual elements of the plot.
}
\examples{
x <- 1:10
y1 <- x^2
y2 <- 2 * x
mplot(x = list(x, x), y = list(y1, y2), main = "Multiple Plots", xlab = "X", ylab = "Y")

}
\references{
For more information, see the documentation of the \code{plot} function in the base package.
}
\seealso{
\code{\link{plot}}
}
\keyword{multiple}
\keyword{plots}
\keyword{plotting}
