% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mpoints}
\alias{mpoints}
\title{mpoints: Multiple Points Plotting Function}
\usage{
mpoints(
  x = NULL,
  y = NULL,
  alp.points = nasd(),
  col.points = nasd(),
  lwd.points = nasd(),
  ...
)
}
\arguments{
\item{x}{A vector or list of x-coordinates.}

\item{y}{A vector or list of y-coordinates.}

\item{alp.points}{The transparency level of the points. Default is NA.}

\item{col.points}{The color(s) of the points. Default is NA.}

\item{lwd.points}{The size(s) of the points. Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create a multiple points plot by specifying the
x and y coordinates. It supports customization of point colors, point
transparency, and point sizes.
}
\examples{
x <- 1:10
y1 <- x
y2 <- 2*x
y3 <- 3*x
mpoints(x, y1, col.points = "red", alp.points = 0.5)
mpoints(x, y2, col.points = "blue", lwd.points = 2)
mpoints(x, y3, col.points = "green", lwd.points = 2)

}
\references{
For more information, see the documentation of the \code{points} function in the base package.
}
\seealso{
\code{\link{points}}
}
\keyword{multiple}
\keyword{plotting}
\keyword{points}
