% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mshadow}
\alias{mshadow}
\title{mshadow}
\usage{
mshadow(
  x1,
  y1 = NULL,
  x2 = NULL,
  y2 = NULL,
  alp.border = nasd(),
  col.border = nasd(),
  lty.border = nasd(),
  lwd.border = nasd(),
  alp.shadow = nasd(),
  col.shadow = nasd(),
  alp.mean = nasd(),
  col.mean = nasd(),
  lwd.mean = nasd(),
  lty.mean = nasd(),
  plot.mean = nasd(),
  plot.shadow = nasd(),
  ...
)
}
\arguments{
\item{x1}{\code{x1}-coordinate.}

\item{y1}{\code{y1}-coordinate.}

\item{x2}{\code{x2}-coordinate.}

\item{y2}{\code{y2}-coordinate.}

\item{alp.border}{border color transparency.}

\item{col.border}{border color.}

\item{lty.border}{border type.}

\item{lwd.border}{border width.}

\item{alp.shadow}{shadow color transparency.}

\item{col.shadow}{shadow color.}

\item{alp.mean}{mean color transparency.}

\item{col.mean}{mean color.}

\item{lwd.mean}{mean width.}

\item{lty.mean}{mean type.}

\item{plot.mean}{to plot mean.}

\item{plot.shadow}{to plot shadow.}
}
\description{
#
}
\examples{
#
}
\seealso{
\code{itz::Shadow} & \code{itz::color}.
}
