% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mshadow}
\alias{mshadow}
\title{mshadow: Multiple Shadow Plot Function}
\usage{
mshadow(
  x1,
  y1 = NULL,
  x2 = NULL,
  y2 = NULL,
  alp.border = nasd(),
  col.border = nasd(),
  lty.border = nasd(),
  lwd.border = nasd(),
  alp.shadow = nasd(),
  col.shadow = nasd(),
  alp.mean = nasd(),
  col.mean = nasd(),
  lwd.mean = nasd(),
  lty.mean = nasd(),
  plot.mean = nasd(),
  plot.shadow = nasd(),
  ...
)
}
\arguments{
\item{x1}{A numeric vector representing the x-axis values for the lower bound of the shadow.}

\item{y1}{A numeric vector representing the y-axis values for the lower bound of the shadow.}

\item{x2}{A numeric vector representing the x-axis values for the upper bound of the shadow.
Default is NULL, in which case x1 will be used as the upper bound.}

\item{y2}{A numeric vector representing the y-axis values for the upper bound of the shadow.
Default is NULL, in which case y1 will be used as the upper bound.}

\item{alp.border}{The transparency level of the shadow border. Default is NA.}

\item{col.border}{The color of the shadow border. Default is NA.}

\item{lty.border}{The line type of the shadow border. Default is NA.}

\item{lwd.border}{The line width of the shadow border. Default is NA.}

\item{alp.shadow}{The transparency level of the shadow. Default is NA.}

\item{col.shadow}{The color of the shadow. Default is NA.}

\item{alp.mean}{The transparency level of the mean line. Default is NA.}

\item{col.mean}{The color of the mean line. Default is NA.}

\item{lwd.mean}{The line width of the mean line. Default is NA.}

\item{lty.mean}{The line type of the mean line. Default is NA.}

\item{plot.mean}{Logical value indicating whether to plot the mean line. Default is NA.}

\item{plot.shadow}{Logical value indicating whether to plot the shadow. Default is NA.}

\item{...}{Additional parameters.}
}
\description{
This function allows you to create multiple shadow plots with customizable settings
such as border color, transparency, line type, line width, shadow transparency,
shadow color, mean transparency, mean color, and plot options. It supports plotting
shadows and mean lines for multiple sets of x-y data.
}
\examples{
x1 <- c(1, 2, 3)
y1 <- c(10, 15, 12)
x2 <- c(1, 2, 3)
y2 <- c(8, 12, 10)
mshadow(x1, y1, x2, y2, col.shadow = "gray", alp.shadow = 0.5, col.mean = "red", alp.mean = 0.8)

}
\seealso{
\code{\link{polygon}}, \code{\link{mlines}}
}
\keyword{multiple}
\keyword{plotting}
\keyword{shadowplot}
